+++ packages/google_mobile_ads/ios/Classes/FLTGoogleMobileAdsReaderWriter_Internal.m
@@ -14,6 +14,11 @@
 
 #import "FLTGoogleMobileAdsReaderWriter_Internal.h"
 #import "FLTAdUtil.h"
+#import "NativeTemplates/FLTNativeTemplateColor.h"
+#import "NativeTemplates/FLTNativeTemplateFontStyle.h"
+#import "NativeTemplates/FLTNativeTemplateStyle.h"
+#import "NativeTemplates/FLTNativeTemplateTextStyle.h"
+#import "NativeTemplates/FLTNativeTemplateType.h"
 
 // The type values below must be consistent for each platform.
 typedef NS_ENUM(NSInteger, FLTAdMobField) {
@@ -36,6 +41,12 @@ typedef NS_ENUM(NSInteger, FLTAdMobField) {
   FLTAdmobFieldVideoOptions = 145,
   FLTAdmobFieldInlineAdaptiveAdSize = 146,
   FLTAdmobRequestConfigurationParams = 148,
+  FLTAdmobFieldNativeTemplateStyle = 149,
+  FLTAdmobFieldNativeTemplateTextStyle = 150,
+  FLTAdmobFieldNativeTemplateFontStyle = 151,
+  FLTAdmobFieldNativeTemplateType = 152,
+  FLTAdmobFieldNativeTemplateColor = 153,
+
 };
 
 @interface FLTGoogleMobileAdsWriter : FlutterStandardWriter
@@ -264,6 +275,60 @@ typedef NS_ENUM(NSInteger, FLTAdMobField) {
               maxHeight:[self readValueOfType:[self readByte]]
             orientation:[self readValueOfType:[self readByte]]];
   }
+  case FLTAdmobFieldNativeTemplateStyle: {
+    FLTNativeTemplateType *templateType =
+        [self readValueOfType:[self readByte]];
+    FLTNativeTemplateColor *mainBackgroundColor =
+        [self readValueOfType:[self readByte]];
+    FLTNativeTemplateTextStyle *callToActionStyle =
+        [self readValueOfType:[self readByte]];
+    FLTNativeTemplateTextStyle *primaryTextStyle =
+        [self readValueOfType:[self readByte]];
+    FLTNativeTemplateTextStyle *secondaryTextStyle =
+        [self readValueOfType:[self readByte]];
+    FLTNativeTemplateTextStyle *tertiaryTextStyle =
+        [self readValueOfType:[self readByte]];
+    NSNumber *cornerRadius = [self readValueOfType:[self readByte]];
+    return
+        [[FLTNativeTemplateStyle alloc] initWithTemplateType:templateType
+                                         mainBackgroundColor:mainBackgroundColor
+                                           callToActionStyle:callToActionStyle
+                                            primaryTextStyle:primaryTextStyle
+                                          secondaryTextStyle:secondaryTextStyle
+                                           tertiaryTextStyle:tertiaryTextStyle
+                                                cornerRadius:cornerRadius];
+  }
+  case FLTAdmobFieldNativeTemplateTextStyle: {
+    FLTNativeTemplateColor *textColor = [self readValueOfType:[self readByte]];
+    FLTNativeTemplateColor *backgroundColor =
+        [self readValueOfType:[self readByte]];
+    FLTNativeTemplateFontStyleWrapper *fontStyle =
+        [self readValueOfType:[self readByte]];
+    NSNumber *size = [self readValueOfType:[self readByte]];
+    return [[FLTNativeTemplateTextStyle alloc] initWithTextColor:textColor
+                                                 backgroundColor:backgroundColor
+                                                       fontStyle:fontStyle
+                                                            size:size];
+  }
+  case FLTAdmobFieldNativeTemplateFontStyle: {
+    NSNumber *fontStyleIndex = [self readValueOfType:[self readByte]];
+    return [[FLTNativeTemplateFontStyleWrapper alloc]
+        initWithInt:fontStyleIndex.intValue];
+  }
+  case FLTAdmobFieldNativeTemplateType: {
+    NSNumber *templateIndex = [self readValueOfType:[self readByte]];
+    return [[FLTNativeTemplateType alloc] initWithInt:templateIndex.intValue];
+  }
+  case FLTAdmobFieldNativeTemplateColor: {
+    NSNumber *alpha = [self readValueOfType:[self readByte]];
+    NSNumber *red = [self readValueOfType:[self readByte]];
+    NSNumber *green = [self readValueOfType:[self readByte]];
+    NSNumber *blue = [self readValueOfType:[self readByte]];
+    return [[FLTNativeTemplateColor alloc] initWithAlpha:alpha
+                                                     red:red
+                                                   green:green
+                                                    blue:blue];
+  }
   }
   return [super readValueOfType:type];
 }
@@ -411,6 +476,38 @@ typedef NS_ENUM(NSInteger, FLTAdMobField) {
     [super writeValue:NSNull.null];
     [super writeValue:NSNull.null];
     [self writeValue:params.testDeviceIdentifiers];
+  } else if ([value isKindOfClass:[FLTNativeTemplateType class]]) {
+    [self writeByte:FLTAdmobFieldNativeTemplateType];
+    FLTNativeTemplateType *templateType = value;
+    [self writeValue:@(templateType.intValue)];
+  } else if ([value isKindOfClass:[FLTNativeTemplateFontStyleWrapper class]]) {
+    [self writeByte:FLTAdmobFieldNativeTemplateFontStyle];
+    FLTNativeTemplateFontStyleWrapper *fontStyle = value;
+    [self writeValue:@(fontStyle.intValue)];
+  } else if ([value isKindOfClass:[FLTNativeTemplateColor class]]) {
+    [self writeByte:FLTAdmobFieldNativeTemplateColor];
+    FLTNativeTemplateColor *templateColor = value;
+    [self writeValue:templateColor.alpha];
+    [self writeValue:templateColor.red];
+    [self writeValue:templateColor.green];
+    [self writeValue:templateColor.blue];
+  } else if ([value isKindOfClass:[FLTNativeTemplateTextStyle class]]) {
+    [self writeByte:FLTAdmobFieldNativeTemplateTextStyle];
+    FLTNativeTemplateTextStyle *textStyle = value;
+    [self writeValue:textStyle.textColor];
+    [self writeValue:textStyle.backgroundColor];
+    [self writeValue:textStyle.fontStyle];
+    [self writeValue:textStyle.size];
+  } else if ([value isKindOfClass:[FLTNativeTemplateStyle class]]) {
+    [self writeByte:FLTAdmobFieldNativeTemplateStyle];
+    FLTNativeTemplateStyle *templateStyle = value;
+    [self writeValue:templateStyle.templateType];
+    [self writeValue:templateStyle.mainBackgroundColor];
+    [self writeValue:templateStyle.callToActionStyle];
+    [self writeValue:templateStyle.primaryTextStyle];
+    [self writeValue:templateStyle.secondaryTextStyle];
+    [self writeValue:templateStyle.tertiaryTextStyle];
+    [self writeValue:templateStyle.cornerRadius];
   } else {
     [super writeValue:value];
   }