+++ packages/google_mobile_ads/CHANGELOG.md
@@ -1,3 +1,12 @@
+## 2.4.0
+* Adds support for native templates, which are predefined layouts for native ads.
+  * `NativeAd` has a new optional parameter, `nativeTemplateStyle` of type `NativeTemplateStyle`.
+    If provided, the plugin will inflate and style a platform native ad view for you, instead of
+    requiring you to write and register a NativeAdFactory (android) or FLTNativeAdFactory (iOS).
+* Adds a new flag, AdWidget.optOutOfVisibilityDetectorWorkaround. Setting this to true
+  lets you opt out of the fix added for https://github.com/googleads/googleads-mobile-flutter/issues/580,
+  which was resolved in Flutter 3.7.0.
+
 ## 2.3.0
 * Updates GMA iOS dependency to 9.13
 * Updates GMA Android dependency to 21.3.0